\name{amf-package}
\alias{amf-package}
\alias{amf}
\docType{package}
\title{
\packageTitle{amf}
}
\description{
\packageDescription{amf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{amf}
\packageIndices{amf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{amf}

Maintainer: \packageMaintainer{amf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
