% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convergencePlot}
\alias{convergencePlot}
\title{Plot model convergence statistics across iterations}
\usage{
convergencePlot(object)
}
\arguments{
\item{object}{valid \code{nmfmodel} object}
}
\value{
ggplot2 plot
}
\description{
Given a model generated with \code{full_path = TRUE}, visualize the decrease in \eqn{h}, \eqn{w} loss and compare against the loss of the model (if applicable).
}
