% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastNMF.R
\name{FastNMF}
\alias{FastNMF}
\title{Run Fast NMF}
\usage{
FastNMF(
  A,
  k,
  tol = 1e-04,
  maxit = 100,
  verbose = FALSE,
  L1 = 0,
  threads = 0,
  mode = 1
)
}
\arguments{
\item{A}{sparse input matrix giving data to be factorized}

\item{k}{rank of the factorization}

\item{tol}{Pearson correlation distance between model across consecutive iterations at which to call convergence and stop updates}

\item{maxit}{maximum number of updates, if \code{tol} remains unsatisfied}

\item{verbose}{print fitting tolerances with each iteration}

\item{L1}{L1/LASSO-like penalty to increase sparsity of the model}

\item{threads}{number of threads to use (0 = let OpenMP decide)}
}
\value{
List of model matrices \code{w}, \code{d}, and \code{h}
}
\description{
See RcppML::nmf for more details
}
\examples{
\dontrun{
data(iris)
model <- FastNMF(as.matrix(iris[,1:4]), 3)
str(model)
heatmap(model$w)
}
}
\author{
Zach DeBruine
}
