% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mdensity}
\alias{mdensity}
\title{Plot the density of maximum sample weights in a model}
\usage{
mdensity(h)
}
\arguments{
\item{h}{model}
}
\value{
ggplot2 object
}
\description{
Useful for determining how well samples map to a feature model in hard clusterings
}
