% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mheatmap}
\alias{mheatmap}
\title{Plot a projection matrix}
\usage{
mheatmap(
  h,
  legend.labels = c(0, 1),
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  show.rownames = "auto",
  show.colnames = "auto",
  Rowv = FALSE,
  Colv = FALSE,
  guide.title = "mapping\\nweight",
  colors = NULL,
  na.color = "#fafafa",
  rownames.text.size = 8,
  discrete = FALSE,
  legend.key.width = 0.2,
  aspect.ratio = 0.5
)
}
\arguments{
\item{h}{matrix to be plotted}

\item{legend.labels}{breakpoints for the continuous color scale legend}

\item{xlab}{x label, or NULL}

\item{ylab}{y label, or NULL}

\item{title}{title for the plot, or NULL}

\item{show.rownames}{show row names}

\item{show.colnames}{show column names}

\item{Rowv}{reorder rows using the hclust ward.d2 method}

\item{Colv}{reorder columns using the hclust ward.d2 method}

\item{guide.title}{title of the colorbar guide}

\item{colors}{array of colors for a gradient (if continuous) or discrete colors (of length corresponding to the number of discrete values in the matrix, or an error will be thrown). If null, an appropriate number of colors will be generated by spinning the ggplot2-style color wheel. An array of numbers may also be specified, corresponding to ggplot color palette indices.}

\item{na.color}{color for NA values}

\item{rownames.text.size}{size of rownames labels}

\item{discrete}{are the values in h discrete (i.e. multinomial). If so, specify discrete = TRUE and a discrete scale will be supplied.}

\item{legend.key.width}{width of legend key in unit "cm"}

\item{aspect.ratio}{theme aspect ratio for the plot (< 1 = wide, > 1 = tall)}
}
\value{
ggplot2 object
}
\description{
ggplot2 heatmap of a projection matrix, ideally only for small matrices. For large matrices, see other packages or submit a rasterized or compressed matrix.
}
\details{
ggplot2 grammar may be applied to add to or overwrite existing settings.
}
