% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\docType{class}
\name{nmfmodel}
\alias{nmfmodel}
\alias{nmfmodel,}
\alias{nmfmodel-class}
\title{The "nmfmodel" class}
\description{
The nmfmodel object holds an nmf factor model and basic statistics computed about the model
}
\section{Slots}{

\describe{
\item{\code{w}}{matrix of \emph{features x factors} of dimensions \emph{m x k}}

\item{\code{d}}{diagonal scaling vector of length \emph{k}}

\item{\code{h}}{matrix of \emph{factors x samples} of dimensions \emph{k x n}}

\item{\code{loss}}{mean squared/absolute error of the model}

\item{\code{tol_wh}}{relative change in \eqn{h} across consecutive iterations multiplied by the 
relative change in \eqn{w} across consecutive iterations.}

\item{\code{tol_loss}}{relative change in loss across consecutive iterations.}

\item{\code{it}}{number of iterations used to solve the model}

\item{\code{path}}{a list of \code{nmfmodel} objects containing intermediate model states along the full
solution path, if they were requested in \code{\link{nmf}}}
}}

\seealso{
\code{\link{nmfmodel}}
}
