% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastnmf-package.R
\docType{package}
\name{fastnmf}
\alias{fastnmf}
\alias{_PACKAGE}
\alias{fastnmf,}
\alias{fastnmf-package}
\title{fastnmf: fast non-negative matrix factorization}
\description{
Fast and adaptive solvers for constrained non-negative least squares and matrix factorization by alternating constrained least squares.
}
\details{
The fastnmf package is built around two functions: \code{nnls} and \code{nmf}.

\code{nnls} is an adaptive solver for non-negative least squares with unprecedented efficiency and flexibility.
\code{nmf} uses the \code{nnls} engine to run non-negative matrix factorization, making use of several new ideas to both increase speed and reduce iterations needed for convergence.
}
\section{Why is fastnmf fast?}{

\itemize{
\item thing 1
\item thing 2
\item thing 3
}
}

\section{Non-negative least squares}{

d
}

\section{Get started!}{

Vignette
}

\section{Advanced tutorials}{

\itemize{
\item Vignette:
\item Vignette:
\item Vignette:
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zdebruine/amf}
  \item Report bugs at \url{https://github.com/zdebruine/amf/issues}
}

}
\author{
Zach DeBruine
}
